Favorite function moving over to redux action
